Updated with all the shaders
GLTF Unlit and MToon10 were also moved here.

These might get outdated overtime, but as long as
the 3D Model Loader is not impacted, it's not a
problem.

Also MToon1.X might be slightly altered to support
normal maps encoded "the Blender way" we'll say,
with main channel being the Blue channel.

Signed-off-by: Voyage <voyage@miouyouyou.fr>